format with prettier

+++ contracts/governance/GovToken/IGovToken.sol
@@ -3,5 +3,8 @@
 pragma solidity ^0.8.0;
 
 interface IGovToken {
-    function getPriorVotes(address account, uint blockNumber) external view returns (uint);
+    function getPriorVotes(address account, uint256 blockNumber)
+        external
+        view
+        returns (uint256);
 }
